+++ src/components/selectImmeuble.jsx
@@ -1,4 +1,3 @@
-import { useNavigate } from 'react-router-dom';
 import { useSelector, useDispatch } from "react-redux";
 import { useEffect, useState } from 'react';
 import { immeubleVisite, setImmeuble } from '../store/visiteSlice.jsx';
@@ -7,7 +6,6 @@ import { getImmeubles } from '../services/api/immeubleApi';
 export default function SelectImmeuble (){
 
     const immeubleVis = useSelector(immeubleVisite);
-    const navigate = useNavigate();
     const dispatch = useDispatch();
     const [listImmeubles, setListImmeubles] = useState([]);
 
@@ -25,7 +23,7 @@ export default function SelectImmeuble (){
         
     return (
         <div className="w-80">
-            {immeubleVis}
+            
             <label htmlFor="countries" className="block mb-2 text-sm font-medium text-gray-900 dark:text-white">Selectionner un immeuble</label>
             <select id="countries" onChange={handleSelect} className="bg-gray-50 border border-gray-300 text-gray-900 text-sm rounded-lg focus:ring-blue-500 focus:border-blue-500 block w-full p-2.5 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-blue-500 dark:focus:border-blue-500">
                 <option defaultValue></option>