+++ src/services/api/immeubleApi.jsx
@@ -1,7 +1,9 @@
 import { api } from "./axiosConfigs.jsx";
 import store from '../../store/store';
 
+// Api call pour récuperer la liste des immeubles
 export async function getImmeubles(){
+    // récupère le token du store
     const authToken = getCurrentStateFromStore().token.value;
     return api.request({
         url: "/immeuble",
@@ -12,6 +14,7 @@ export async function getImmeubles(){
     })
 }
 
+// Permet de récupérer les states du store 
 const getCurrentStateFromStore = () =>{
     return {
         token: store.getState().token,