+++ src/views/login.jsx
@@ -6,7 +6,7 @@ export default function Login() {
     <div className="h-screen flex items-center">
       <div className="m-auto">
         <div className="flex justify-center pb-9">
-          <h1 className="text-white md:text-4xl text-3xl">Visite d'immeuble .NET</h1>
+          <h1 className="text-white text-3xl">Bienvenu sur NaviLite</h1>
         </div>
         <LoginForm></LoginForm>
       </div>