+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 == Changelog ==
 
-0.4
+0.7 - 24/02/2023
+
+added
+
+- Ajout de persist react pour stocker les states
+- Extraction des appels api dans un service
+- Création de la page accueil
+
+0.4 - 22/02/2023
 
 added